refactor: spot trade order form
