+++ components/partials/spot/trading/order-details.vue
@@ -7,7 +7,10 @@
     >
       <div class="mt-4">
         <TextInfo :title="$t('trade.price')" class="mt-2">
-          <span v-if="price.gt(0)" class="font-mono flex items-start break-all">
+          <span
+            v-if="executionPrice.gt(0)"
+            class="font-mono flex items-start break-all"
+          >
             {{ priceToFormat }}
             <span class="text-gray-500 ml-1 break-normal">
               {{ market.quoteToken.symbol }}
@@ -135,12 +138,14 @@
 
 <script lang="ts">
 import Vue, { PropType } from 'vue'
+import { cosmosSdkDecToBigNumber } from '@injectivelabs/sdk-ts'
 import { BigNumberInBase } from '@injectivelabs/utils'
 import {
   UiSpotMarketWithToken,
   ZERO_IN_BASE,
   SpotOrderSide
 } from '@injectivelabs/sdk-ui-ts'
+import { TradingRewardsCampaign } from '~/app/client/types/exchange'
 import Drawer from '~/components/elements/drawer.vue'
 import { Icon } from '~/types'
 import { UI_DEFAULT_PRICE_DISPLAY_DECIMALS } from '~/app/utils/constants'
@@ -172,16 +177,6 @@ export default Vue.extend({
       type: Object as PropType<BigNumberInBase>
     },
 
-    feeReturned: {
-      required: true,
-      type: Object as PropType<BigNumberInBase>
-    },
-
-    feeRebates: {
-      required: true,
-      type: Object as PropType<BigNumberInBase>
-    },
-
     takerFeeRateDiscount: {
       required: true,
       type: Object as PropType<BigNumberInBase>
@@ -202,22 +197,12 @@ export default Vue.extend({
       type: Object as PropType<BigNumberInBase>
     },
 
-    takerExpectedPts: {
-      required: true,
-      type: Object as PropType<BigNumberInBase>
-    },
-
-    makerExpectedPts: {
-      required: true,
-      type: Object as PropType<BigNumberInBase>
-    },
-
     fees: {
       required: true,
       type: Object as PropType<BigNumberInBase>
     },
 
-    price: {
+    executionPrice: {
       required: true,
       type: Object as PropType<BigNumberInBase>
     },
@@ -245,11 +230,6 @@ export default Vue.extend({
     feeRate: {
       required: true,
       type: Object as PropType<BigNumberInBase>
-    },
-
-    executionPrice: {
-      required: true,
-      type: Object as PropType<BigNumberInBase>
     }
   },
 
@@ -262,6 +242,170 @@ export default Vue.extend({
   computed: {
     market(): UiSpotMarketWithToken | undefined {
       return this.$accessor.spot.market
+    },
+
+    tradingRewardsCampaign(): TradingRewardsCampaign | undefined {
+      return this.$accessor.exchange.tradingRewardsCampaign
+    },
+
+    totalWithoutFees(): BigNumberInBase {
+      const { fees, total, market } = this
+
+      if (total.isNaN() || total.lte(0) || !market) {
+        return ZERO_IN_BASE
+      }
+
+      return total.minus(fees)
+    },
+
+    feeReturned(): BigNumberInBase {
+      const { total, takerFeeRate, makerFeeRate, market } = this
+
+      if (total.isNaN() || total.lte(0) || !market) {
+        return ZERO_IN_BASE
+      }
+
+      return total.times(
+        new BigNumberInBase(takerFeeRate).minus(makerFeeRate.abs())
+      )
+    },
+
+    feeRebates(): BigNumberInBase {
+      const { total, makerFeeRate, market } = this
+
+      if (total.isNaN() || !market) {
+        return ZERO_IN_BASE
+      }
+
+      return new BigNumberInBase(total.times(makerFeeRate).abs()).times(
+        0.6 /* Only 60% of the fees are getting returned */
+      )
+    },
+
+    hubUrl(): string {
+      return 'https://hub.injective.network/bridge'
+    },
+
+    makerExpectedPts(): BigNumberInBase {
+      const { market, makerFeeRate, tradingRewardsCampaign, fees } = this
+
+      if (!market) {
+        return ZERO_IN_BASE
+      }
+
+      if (makerFeeRate.lte(0)) {
+        return ZERO_IN_BASE
+      }
+
+      if (!tradingRewardsCampaign) {
+        return ZERO_IN_BASE
+      }
+
+      if (!tradingRewardsCampaign.tradingRewardCampaignInfo) {
+        return ZERO_IN_BASE
+      }
+
+      const disqualified =
+        tradingRewardsCampaign.tradingRewardCampaignInfo.disqualifiedMarketIdsList.find(
+          (marketId) => marketId === market.marketId
+        )
+
+      if (disqualified) {
+        return ZERO_IN_BASE
+      }
+
+      const denomIncluded =
+        tradingRewardsCampaign.tradingRewardCampaignInfo.quoteDenomsList.find(
+          (denom) => denom === market.quoteDenom
+        )
+
+      if (!denomIncluded) {
+        return ZERO_IN_BASE
+      }
+
+      const boostedList = tradingRewardsCampaign.tradingRewardCampaignInfo
+        .tradingRewardBoostInfo
+        ? tradingRewardsCampaign.tradingRewardCampaignInfo
+            .tradingRewardBoostInfo.boostedSpotMarketIdsList
+        : []
+      const multipliersList = tradingRewardsCampaign.tradingRewardCampaignInfo
+        .tradingRewardBoostInfo
+        ? tradingRewardsCampaign.tradingRewardCampaignInfo
+            .tradingRewardBoostInfo.spotMarketMultipliersList
+        : []
+
+      const boosted = boostedList.findIndex(
+        (spotMarketId) => spotMarketId === market.marketId
+      )
+      const boostedMultiplier =
+        boosted >= 0
+          ? cosmosSdkDecToBigNumber(
+              multipliersList[boosted]
+                ? multipliersList[boosted].makerPointsMultiplier
+                : 1
+            )
+          : 1
+
+      return new BigNumberInBase(fees).times(boostedMultiplier)
+    },
+
+    takerExpectedPts(): BigNumberInBase {
+      const { market, tradingRewardsCampaign, fees } = this
+
+      if (!market) {
+        return ZERO_IN_BASE
+      }
+
+      if (!tradingRewardsCampaign) {
+        return ZERO_IN_BASE
+      }
+
+      if (!tradingRewardsCampaign.tradingRewardCampaignInfo) {
+        return ZERO_IN_BASE
+      }
+
+      const disqualified =
+        tradingRewardsCampaign.tradingRewardCampaignInfo.disqualifiedMarketIdsList.find(
+          (marketId) => marketId === market.marketId
+        )
+
+      if (disqualified) {
+        return ZERO_IN_BASE
+      }
+
+      const denomIncluded =
+        tradingRewardsCampaign.tradingRewardCampaignInfo.quoteDenomsList.find(
+          (denom) => denom === market.quoteDenom
+        )
+
+      if (!denomIncluded) {
+        return ZERO_IN_BASE
+      }
+
+      const boostedList = tradingRewardsCampaign.tradingRewardCampaignInfo
+        .tradingRewardBoostInfo
+        ? tradingRewardsCampaign.tradingRewardCampaignInfo
+            .tradingRewardBoostInfo.boostedSpotMarketIdsList
+        : []
+      const multipliersList = tradingRewardsCampaign.tradingRewardCampaignInfo
+        .tradingRewardBoostInfo
+        ? tradingRewardsCampaign.tradingRewardCampaignInfo
+            .tradingRewardBoostInfo.spotMarketMultipliersList
+        : []
+
+      const boosted = boostedList.findIndex(
+        (spotMarketId) => spotMarketId === market.marketId
+      )
+      const boostedMultiplier =
+        boosted >= 0
+          ? cosmosSdkDecToBigNumber(
+              multipliersList[boosted]
+                ? multipliersList[boosted].takerPointsMultiplier
+                : 1
+            )
+          : 1
+
+      return new BigNumberInBase(fees).times(boostedMultiplier)
     },
 
     extractedTotal(): BigNumberInBase {
@@ -285,9 +429,9 @@ export default Vue.extend({
     },
 
     totalEstimatedFees(): string {
-      const { price, amount, feeRate, market } = this
+      const { executionPrice, amount, feeRate, market } = this
 
-      const fees = price.times(amount).times(feeRate)
+      const fees = executionPrice.times(amount).times(feeRate)
 
       if (!market) {
         return fees.toFormat(UI_DEFAULT_PRICE_DISPLAY_DECIMALS)
@@ -297,13 +441,13 @@ export default Vue.extend({
     },
 
     priceToFormat(): string {
-      const { price, market } = this
+      const { executionPrice, market } = this
 
       if (!market) {
-        return price.toFormat(UI_DEFAULT_PRICE_DISPLAY_DECIMALS)
+        return executionPrice.toFormat(UI_DEFAULT_PRICE_DISPLAY_DECIMALS)
       }
 
-      return price.toFormat(market.priceDecimals)
+      return executionPrice.toFormat(market.priceDecimals)
     },
 
     feesToFormat(): string {
@@ -379,10 +523,10 @@ export default Vue.extend({
       const {
         amount,
         market,
+        executionPrice,
         orderTypeBuy,
         quoteAmount,
-        feeRate,
-        executionPrice
+        feeRate
       } = this
 
       if (quoteAmount.isNaN()) {
